--story=1002973 --user=林邦泉 用户增加uuid字段，角色增加uuid字段 https://www.tapd.cn/54247054/s/1065587

+++ src/main/java/codedriver/module/tenant/api/role/RoleUserSaveApi.java
@@ -2,7 +2,10 @@ package codedriver.module.tenant.api.role;
 
 import codedriver.framework.apiparam.core.ApiParamType;
 import codedriver.framework.dao.mapper.RoleMapper;
+import codedriver.framework.dao.mapper.UserMapper;
 import codedriver.framework.dto.UserVo;
+import codedriver.framework.exception.role.RoleNotFoundException;
+import codedriver.framework.exception.user.UserNotFoundException;
 import codedriver.framework.restful.annotation.Description;
 import codedriver.framework.restful.annotation.Input;
 import codedriver.framework.restful.annotation.Param;
@@ -24,6 +27,9 @@ public class RoleUserSaveApi extends ApiComponentBase {
 
     @Autowired
     private RoleMapper roleMapper;
+    
+    @Autowired
+    private UserMapper userMapper;
 
     @Override
     public String getToken() {
@@ -53,12 +59,18 @@ public class RoleUserSaveApi extends ApiComponentBase {
     @Override
     public Object myDoService(JSONObject jsonObj) throws Exception {
         String roleUuid = jsonObj.getString("roleUuid");
+        if(roleMapper.checkRoleIsExists(roleUuid) == 0) {
+			throw new RoleNotFoundException(roleUuid);
+		}
         roleMapper.deleteUserRoleByRoleUuid(roleUuid);
 		List<String> userUuidList = JSON.parseArray(jsonObj.getString("userUuidList"), String.class);
 		if (CollectionUtils.isNotEmpty(userUuidList)){
 			UserVo userVo = new UserVo();
 			userVo.setRoleUuid(roleUuid);
 			for (String userUuid : userUuidList){
+				if(userMapper.checkUserIsExists(userUuid) == 0) {
+					throw new UserNotFoundException(userUuid);
+				}
 				userVo.setUuid(userUuid);
 				roleMapper.insertRoleUser(userVo);
 			}
